requirements.txt
asgiref==3.8.1
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
whitenoise==5.2.0
pylint==3.1.0